添加Okhttp 实现

+++ WX-App/src/main/java/com/wx/download/WXDownLoadViewModel.kt
@@ -86,7 +86,7 @@ class WXDownLoadViewModel : ViewModel() {
                 }
 
                 is WXState.Downloading -> {
-                    _datas.value!![it.which].progress.value = it.progress.toFloat()
+                    _datas.value!![it.which].progress.value = it.progress
                 }
 
                 is WXState.Pause -> {}
